+++ rana_qgis_plugin/workers.py
@@ -3,10 +3,12 @@ import io
 import json
 import math
 import os
+import shutil
 import zipfile
+from functools import cached_property
 from pathlib import Path
 from time import sleep
-from typing import List
+from typing import List, Optional
 
 import requests
 from bridgestyle.mapboxgl.fromgeostyler import convertGroup
@@ -29,11 +31,13 @@ from rana_qgis_plugin.widgets.utils_avatars import get_avatar
 from .utils import (
     build_vrt,
     get_local_file_path,
+    get_publication_layer_path,
     image_to_bytes,
     split_scenario_extent,
 )
 from .utils_api import (
     finish_file_upload,
+    get_publication_style,
     get_raster_file_link,
     get_raster_style_file,
     get_raster_style_upload_urls,
@@ -51,60 +55,187 @@ from .utils_api import (
 CHUNK_SIZE = 1024 * 1024  # 1 MB
 
 
-class FileDownloadWorker(QThread):
-    """Worker thread for downloading files."""
+class FileDownloadWorkerSignals(QObject):
+    progress = pyqtSignal(
+        int, str
+    )  # Progress signal: emits progress percentage and additional info
+    finished = pyqtSignal(
+        dict, dict, str
+    )  # Finished signal: emits project, file, and local file path
+    failed = pyqtSignal(str)  # Failed signal: emits error messages
+    all_finished = pyqtSignal()
 
-    progress = pyqtSignal(int, str)
-    finished = pyqtSignal(dict, dict, str)
-    failed = pyqtSignal(str)
 
+class FileDownloadBase:
+    """Base class that handles the common download logic."""
+
+    # Note: The signals are removed from FileDownloadBase as they will now exist in the worker classes.
+    def __init__(self, project, file):
+        self.project = project
+        self.file = file
+
+    def get_local_file_path(self) -> tuple[str, str]:
+        raise NotImplementedError
+
+    def get_style_zip(self):
+        raise NotImplementedError
+
+    def download_file(self, signals: FileDownloadWorkerSignals, download_file=True):
+        """Handles the core logic for downloading a file and emits signals from the worker."""
+        path = self.file["id"]
+        descriptor_id = self.file["descriptor_id"]
+        url = get_tenant_file_url(self.project["id"], {"path": path})
+        local_dir_structure, local_file_path = self.get_local_file_path()
+        os.makedirs(local_dir_structure, exist_ok=True)
+        try:
+            if download_file:
+                with requests.get(url, stream=True) as response:
+                    response.raise_for_status()
+                    total_size = int(response.headers.get("content-length", 0))
+                    downloaded_size = 0
+                    previous_progress = -1
+                    with open(local_file_path, "wb") as local_file:
+                        for chunk in response.iter_content(chunk_size=CHUNK_SIZE):
+                            local_file.write(chunk)
+                            downloaded_size += len(chunk)
+                            progress = int((downloaded_size / total_size) * 100)
+                            if progress > previous_progress:
+                                signals.progress.emit(progress, local_file_path)
+                                previous_progress = progress
+            # Handle QML files for vector and raster data
+            self._handle_qml_extraction(descriptor_id, local_dir_structure)
+            # Emit finished signal from the worker
+            signals.finished.emit(self.project, self.file, local_file_path)
+        except requests.exceptions.RequestException as e:
+            signals.failed.emit(f"Failed to download file: {str(e)}")
+        except Exception as e:
+            signals.failed.emit(f"An error occurred: {str(e)}")
+
+    def _handle_qml_extraction(self, descriptor_id, local_dir_structure):
+        """Handles the extraction of QML zip file if required."""
+        if self.file["data_type"] in ["vector", "raster"]:
+            qml_zip_content = self.get_style_zip()
+            if qml_zip_content:
+                stream = io.BytesIO(qml_zip_content)
+                if zipfile.is_zipfile(stream):
+                    with zipfile.ZipFile(stream, "r") as zip_file:
+                        zip_file.extractall(local_dir_structure)
+
+
+class FileDownloadForFileTree(FileDownloadBase):
+    def get_local_file_path(self) -> tuple[str, str]:
+        return get_local_file_path(self.project["slug"], self.file["id"])
+
+    def get_style_zip(self):
+        if self.file["data_type"] == "raster":
+            return get_raster_style_file(self.file["descriptor_id"], "qml.zip")
+        else:
+            return get_vector_style_file(self.file["descriptor_id"], "qml.zip")
+
+
+class FileDownloadForPublicationTree(FileDownloadBase):
     def __init__(
         self,
         project: dict,
         file: dict,
+        publication_id: str,
+        publication_tree: list[str],
+        publication_version: int,
+        style_id: Optional[str] = None,
+        layer_in_file: Optional[str] = None,
     ):
+        super().__init__(project, file)
+        self.publication_tree = publication_tree
+        self.publication_id = publication_id
+        self.publication_version = publication_version
+        self.layer_in_file = layer_in_file
+        self.style_id = style_id
+
+    def get_local_file_path(self) -> tuple[str, str]:
+        return get_publication_layer_path(
+            self.project["slug"], self.file["id"], self.publication_tree
+        )
+
+    def get_style_zip(self):
+        if self.style_id and self.file["data_type"] in ["raster", "vector"]:
+            style_zip = get_publication_style(
+                self.publication_id,
+                self.style_id,
+                self.publication_version,
+                "qml.zip",
+            )
+            return style_zip
+
+
+class SingleFileDownloadWorker(QThread):
+    """Worker thread for downloading a single file."""
+
+    def __init__(self, downloader: FileDownloadBase):
         super().__init__()
-        self.project = project
-        self.file = file
+        self.signals = FileDownloadWorkerSignals()
+        self.downloader = downloader
 
     @pyqtSlot()
     def run(self):
-        project_slug = self.project["slug"]
-        path = self.file["id"]
-        descriptor_id = self.file["descriptor_id"]
-        url = get_tenant_file_url(self.project["id"], {"path": path})
-        local_dir_structure, local_file_path = get_local_file_path(project_slug, path)
-        os.makedirs(local_dir_structure, exist_ok=True)
-        try:
-            with requests.get(url, stream=True) as response:
-                response.raise_for_status()
-                total_size = int(response.headers.get("content-length", 0))
-                downloaded_size = 0
-                previous_progress = -1
-                with open(local_file_path, "wb") as file:
-                    for chunk in response.iter_content(chunk_size=CHUNK_SIZE):
-                        file.write(chunk)
-                        downloaded_size += len(chunk)
-                        progress = int((downloaded_size / total_size) * 100)
-                        if progress > previous_progress:
-                            self.progress.emit(progress, "")
-                            previous_progress = progress
-            # Fetch and extract the QML zip for vector files
-            if self.file["data_type"] in ["vector", "raster"]:
-                if self.file["data_type"] == "raster":
-                    qml_zip_content = get_raster_style_file(descriptor_id, "qml.zip")
-                else:
-                    qml_zip_content = get_vector_style_file(descriptor_id, "qml.zip")
-                if qml_zip_content:
-                    stream = io.BytesIO(qml_zip_content)
-                    if zipfile.is_zipfile(stream):
-                        with zipfile.ZipFile(stream, "r") as zip_file:
-                            zip_file.extractall(local_dir_structure)
-            self.finished.emit(self.project, self.file, local_file_path)
-        except requests.exceptions.RequestException as e:
-            self.failed.emit(f"Failed to download file: {str(e)}")
-        except Exception as e:
-            self.failed.emit(f"An error occurred: {str(e)}")
+        self.downloader.download_file(self.signals)
+
+
+class BatchFileDownloadWorker(QThread):
+    """Worker thread for downloading multiple files, one after the other."""
+
+    def __init__(self, downloaders: list[FileDownloadBase]):
+        super().__init__()
+        self.signals = FileDownloadWorkerSignals()
+        self.downloaders = downloaders
+        self.downloaded_files = {str: str}
+
+    @cached_property
+    def unique_file_ids(self):
+        return set([downloader.file["id"] for downloader in self.downloaders])
+
+    @property
+    def nof_files(self) -> int:
+        """Count number of unique files"""
+        return len(self.unique_file_ids)
+
+    def handle_existing(self, downloader) -> bool:
+        """Check if a file was already downloaded by this worker. If so, just copy the file to the required destination"""
+        if downloader.file["id"] in self.downloaded_files:
+            _, download_path = downloader.get_local_file_path()
+            file_location = self.downloaded_files[downloader.file["id"]]
+            # make sure the file didn't disappear somehow
+            if file_location == download_path:
+                return True
+            if Path(file_location).exists():
+                try:
+                    Path(download_path).parent.mkdir(parents=True, exist_ok=True)
+                    shutil.copy2(file_location, download_path)
+                    return True
+                except (FileNotFoundError, PermissionError, OSError) as e:
+                    # Don't raise for reasonable exceptions, just return False to redownload
+                    return False
+        else:
+            return False
+
+    @pyqtSlot()
+    def run(self):
+        # Find the first downloader for each unique file id, download the file and remove the downloader
+        for file_id in self.unique_file_ids:
+            downloader = next(
+                downloader
+                for downloader in self.downloaders
+                if downloader.file["id"] == file_id
+            )
+            downloader.download_file(self.signals)
+            _, download_path = downloader.get_local_file_path()
+            self.downloaded_files[downloader.file["id"]] = download_path
+            self.downloaders.remove(downloader)
+        # Iterate over the remaining downloaders and copy the existing file
+        for downloader in self.downloaders:
+            # copy existing, and if redownload if that was unsuccessful
+            download_file = not self.handle_existing(downloader)
+            downloader.download_file(self.signals, download_file)
+        self.signals.all_finished.emit()
 
 
 class FileUploadWorker(QThread):